+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Breaking changes
 - Rename `Subscribable` to `Listen`
 - `Emitter#emit()` and `#emitAsync()` are renamed to `#emitSync()` and `#emit()`
+- Rename `Disposable` to `Subscription` and `#dispose()` to `#unsubscribe()` etc.
 
 ### Features
 - The second argument of `Emitter#emit()` and `#emitSync()` (previously `#emitAsync()` and `#emit()`) are now optional if its type is `undefined`